organizing JS
