+++ app/javascript/packs/application.js
@@ -1,5 +1,6 @@
 import "bootstrap";
 
 import "../plugins/flatpickr"
-
-
+import "../plugins/flipback"
+import "../plugins/flipfront"
+import "../plugins/recipe_dropdown"
