+++ app/javascript/plugins/flatpickr.js
@@ -1,4 +1,6 @@
 import flatpickr from "flatpickr"
 import "flatpickr/dist/flatpickr.min.css" // Note this is important!
 
-flatpickr(".datepicker", {})
+flatpickr(".datepicker", {
+  altInput: true
+})
