+++ app/views/deliveries/new.html.erb
@@ -11,7 +11,6 @@
       </div>
       <div class="col-xs-12">
         <%= f.input :address, placeholder: 'Rue du Port-Franc 22, Lausanne' %>
-        <%= @delivery.user.address %>
       </div>
     </div>
     <div class="col-xs-8 col-xs-offset-2">
