+++ app/views/layouts/application.html.erb
@@ -19,9 +19,6 @@
       <%= render 'shared/footer' %>
       <%= javascript_include_tag 'application' %>
       <%= javascript_pack_tag 'application' %>
-      <%= javascript_pack_tag 'flipback' %>
-      <%= javascript_pack_tag 'flipfront' %>
-      <%= javascript_pack_tag 'recipe_dropdown' %>
     </div>
   </body>
 </html>
